+++ src/aircheck/core/checks.py
@@ -8,6 +8,8 @@ __all__ = (
 
 from typing import TYPE_CHECKING, NamedTuple
 
+from aircheck.core.utils import concat_errors
+
 if TYPE_CHECKING:
     from airflow.models import DAG
 
@@ -19,41 +21,47 @@ class CheckResult(NamedTuple):
 
 def check_for_duplicated_dags(dag_ids: list[str]) -> CheckResult:
     seen = set()
+    errors = []
 
     for dag in dag_ids:
         if dag in seen:
-            return CheckResult(False, err_msg=f"DAG '{dag}' has duplicates")
+            errors.append(f"DAG '{dag}' has duplicates")
 
         seen.add(dag)
+
+    if errors:
+        return CheckResult(check_successful=False, err_msg=concat_errors(errors))
 
     return CheckResult(check_successful=True)
 
 
-def check_dag_id_prefix(dag_ids: list[str], expected_prefix: str) -> CheckResult:
-    for dag in dag_ids:
-        if not dag.startswith(expected_prefix):
-            return CheckResult(
-                False,
-                err_msg=f"DAG '{dag}' does not include required prefix {expected_prefix}",
-            )
+def check_dag_id_prefix(dag: "DAG", expected_prefix: str) -> CheckResult:
+    if not dag.dag_id.startswith(expected_prefix):
+        msg = f"DAG '{dag.dag_id}' does not include required prefix '{expected_prefix}'"
+        return CheckResult(
+            False,
+            err_msg=f"{dag.fileloc}: {msg}",
+        )
 
     return CheckResult(check_successful=True)
 
 
 def check_for_empty_dag(dag: "DAG") -> CheckResult:
     if not dag.tasks:
-        return CheckResult(
-            False, err_msg=f"DAG '{dag.dag_id}' must have at least one task"
-        )
+        msg = f"DAG '{dag.dag_id}' must have at least one task"
+        return CheckResult(False, err_msg=f"{dag.fileloc}: {msg}")
     return CheckResult(check_successful=True)
 
 
 def check_for_dangling_tasks(dag: "DAG") -> CheckResult:
     for task in dag.tasks:
         if not task.upstream_list and not task.downstream_list:
+            msg = f"Dangling task '{task.task_id}' in DAG '{dag.dag_id}' (no upstream or downstream dependencies)"
+            err_msg = f"{dag.fileloc}: {msg}"
+
             return CheckResult(
                 False,
-                err_msg=f"Dangling task '{task.task_id}' in DAG '{dag.dag_id}' (no upstream or downstream dependencies)",
+                err_msg=err_msg,
             )
 
     return CheckResult(check_successful=True)