+++ src/aircheck/integrity_check.py
@@ -8,7 +8,7 @@ from aircheck.core.checks import (
     check_for_empty_dag,
 )
 from aircheck.core.load import load_dags
-from aircheck.core.utils import get_dag_modules
+from aircheck.core.utils import concat_errors, get_dag_modules
 
 
 def check_dags_integrity(
@@ -18,38 +18,44 @@ def check_dags_integrity(
     check_empty_dags: bool,
     check_dangling_tasks: bool,
 ) -> CheckResult:
-    dag_modules = get_dag_modules(dag_path, files)
+    if files:
+        dag_modules = get_dag_modules(dag_path, files)
 
-    del files  # no need to keep them anymore and in case of `pre-commit run --all-files` this could get big
+        del files
 
-    if not dag_modules:
-        # no changes made to DAGs - no need to run integrity check
-        return CheckResult(check_successful=True)
+        if not dag_modules:
+            # no changes made to DAGs - no need to run integrity check
+            return CheckResult(check_successful=True)
+
+    errors: list[str] = []
 
     dag_info = load_dags(dag_path=dag_path)
     if dag_info.import_errors:
         errors = [f"{err.file}: {err.msg}" for err in dag_info.import_errors]
-        err_msg = """\n""".join(errors)
+        err_msg = concat_errors(errors)
         return CheckResult(False, err_msg=err_msg)
 
     result = check_for_duplicated_dags(dag_info.dag_ids)
     if not result.check_successful:
-        return result
-
-    if dag_id_prefix:
-        result = check_dag_id_prefix(dag_info.dag_ids, dag_id_prefix)
-        if not result.check_successful:
-            return result
+        errors.append(result.err_msg)
 
     for dag in dag_info.dags:
+        if dag_id_prefix:
+            result = check_dag_id_prefix(dag, dag_id_prefix)
+            if not result.check_successful:
+                errors.append(result.err_msg)
+
         if check_empty_dags:
             result = check_for_empty_dag(dag)
             if not result.check_successful:
-                return result
+                errors.append(result.err_msg)
 
         if check_dangling_tasks:
             result = check_for_dangling_tasks(dag)
             if not result.check_successful:
-                return result
+                errors.append(result.err_msg)
+
+    if errors:
+        result = CheckResult(check_successful=False, err_msg=concat_errors(errors))
 
     return result